Fixed iiif url for thumbnails with specific size Parse iddocs as long, not as int removed contentServer and contentServerWrapper services from web.xml

+++ goobi-viewer-core/src/de/intranda/digiverso/presentation/model/bookshelf/BookshelfItem.java
@@ -267,7 +267,7 @@ public class BookshelfItem implements Serializable {
             
             
             ThumbnailHandler thumbs = BeanUtils.getImageDeliveryBean().getThumb();
-            StructElement doc = new StructElement(Integer.parseInt(luceneId), docs.get(0));
+            StructElement doc = new StructElement(Long.parseLong(luceneId), docs.get(0));
             return thumbs.getThumbnailUrl(doc, width, height);
         } else {
             return "";
